+++ plugin/src/main/java/com/stratio/cassandra/lucene/search/condition/BitemporalCondition.java
@@ -18,16 +18,15 @@
 
 package com.stratio.cassandra.lucene.search.condition;
 
+import com.stratio.cassandra.lucene.schema.mapping.BitemporalMapper;
+import org.apache.lucene.analysis.Analyzer;
+import org.apache.lucene.search.BooleanQuery;
+import org.apache.lucene.search.Query;
+
 import static com.stratio.cassandra.lucene.schema.mapping.BitemporalMapper.BitemporalDateTime;
 import static org.apache.lucene.search.BooleanClause.Occur.MUST;
 import static org.apache.lucene.search.BooleanClause.Occur.SHOULD;
-
-import org.apache.lucene.analysis.Analyzer;
-import org.apache.lucene.search.BooleanQuery;
-import org.apache.lucene.search.NumericRangeQuery;
-import org.apache.lucene.search.Query;
-
-import com.stratio.cassandra.lucene.schema.mapping.BitemporalMapper;
+import static org.apache.lucene.search.NumericRangeQuery.newLongRange;
 
 /**
  * A {@link Condition} implementation that matches bi-temporal (four) fields within two range of values.
@@ -58,21 +57,15 @@ public class BitemporalCondition extends SingleMapperCondition<BitemporalMapper>
      * Constructs a query selecting all fields that intersects with valid time and transaction time ranges including
      * limits.
      *
-     * @param boost     The boost for this query clause. Documents matching this clause will (in addition to the normal
-     *                  weightings) have their score multiplied by {@code boost}.
-     * @param field     The name of the field to be matched.
-     * @param vtFrom    The Valid Time Start.
-     * @param vtTo      The Valid Time End.
-     * @param ttFrom    The Transaction Time Start.
-     * @param ttTo      The Transaction Time End.
-     *
+     * @param boost  The boost for this query clause. Documents matching this clause will (in addition to the normal
+     *               weightings) have their score multiplied by {@code boost}.
+     * @param field  The name of the field to be matched.
+     * @param vtFrom The Valid Time Start.
+     * @param vtTo   The Valid Time End.
+     * @param ttFrom The Transaction Time Start.
+     * @param ttTo   The Transaction Time End.
      */
-    public BitemporalCondition(Float boost,
-                               String field,
-                               Object vtFrom,
-                               Object vtTo,
-                               Object ttFrom,
-                               Object ttTo) {
+    public BitemporalCondition(Float boost, String field, Object vtFrom, Object vtTo, Object ttFrom, Object ttTo) {
         super(boost, field, BitemporalMapper.class);
         this.vtFrom = vtFrom;
         this.vtTo = vtTo;
@@ -80,52 +73,50 @@ public class BitemporalCondition extends SingleMapperCondition<BitemporalMapper>
         this.ttTo = ttTo;
     }
 
+    /** {@inheritDoc} */
     @Override
     public Query query(BitemporalMapper mapper, Analyzer analyzer) {
 
-        Long vt_from = this.vtFrom == null ?
-                new BitemporalDateTime(DEFAULT_FROM).toDate().getTime() :
-                mapper.parseBitemporalDate(this.vtFrom).toDate().getTime();
-        Long vt_to = this.vtTo == null ?
-                new BitemporalDateTime(DEFAULT_TO).toDate().getTime() :
-                mapper.parseBitemporalDate(this.vtTo).toDate().getTime();
-        Long tt_from = this.ttFrom == null ?
-                new BitemporalDateTime(DEFAULT_FROM).toDate().getTime() :
-                mapper.parseBitemporalDate(this.ttFrom).toDate().getTime();
-        Long tt_to = this.ttTo == null ?
-                new BitemporalDateTime(DEFAULT_TO).toDate().getTime() :
-                mapper.parseBitemporalDate(this.ttTo).toDate().getTime();
+        Long vtFromTime = parseTime(mapper, DEFAULT_FROM, vtFrom);
+        Long vtToTime = parseTime(mapper, DEFAULT_TO, vtTo);
+        Long ttFromTime = parseTime(mapper, DEFAULT_FROM, ttFrom);
+        Long ttToTime = parseTime(mapper, DEFAULT_TO, ttTo);
 
-        Long MIN=BitemporalDateTime.MIN.toDate().getTime();
-        Long MAX=BitemporalDateTime.MAX.toDate().getTime();
+        Long minTime = BitemporalDateTime.MIN.toDate().getTime();
+        Long maxTime = BitemporalDateTime.MAX.toDate().getTime();
 
         BooleanQuery.Builder builder = new BooleanQuery.Builder();
 
-        BooleanQuery.Builder validBuilder= new BooleanQuery.Builder();
-        validBuilder.add(NumericRangeQuery.newLongRange(field + ".vtFrom", vt_from, vt_to, true, true), SHOULD);
-        validBuilder.add(NumericRangeQuery.newLongRange(field + ".vtTo", vt_from, vt_to, true, true),SHOULD);
+        BooleanQuery.Builder validBuilder = new BooleanQuery.Builder();
+        validBuilder.add(newLongRange(field + ".vtFrom", vtFromTime, vtToTime, true, true), SHOULD);
+        validBuilder.add(newLongRange(field + ".vtTo", vtFromTime, vtToTime, true, true), SHOULD);
 
-        BooleanQuery.Builder containsValidBuilder= new BooleanQuery.Builder();
-        containsValidBuilder.add(NumericRangeQuery.newLongRange(field+".vtFrom",MIN,vt_from,true,true),MUST);
-        containsValidBuilder.add(NumericRangeQuery.newLongRange(field+".vtTo ",vt_to,MAX,true,true),MUST);
-        validBuilder.add(containsValidBuilder.build(),SHOULD);
+        BooleanQuery.Builder containsValidBuilder = new BooleanQuery.Builder();
+        containsValidBuilder.add(newLongRange(field + ".vtFrom", minTime, vtFromTime, true, true), MUST);
+        containsValidBuilder.add(newLongRange(field + ".vtTo ", vtToTime, maxTime, true, true), MUST);
+        validBuilder.add(containsValidBuilder.build(), SHOULD);
 
+        BooleanQuery.Builder transactionBuilder = new BooleanQuery.Builder();
+        transactionBuilder.add(newLongRange(field + ".ttFrom", ttFromTime, ttToTime, true, true), SHOULD);
+        transactionBuilder.add(newLongRange(field + ".ttTo", ttFromTime, ttToTime, true, true), SHOULD);
 
-        BooleanQuery.Builder transactionBuilder= new BooleanQuery.Builder();
-        transactionBuilder.add(NumericRangeQuery.newLongRange(field + ".ttFrom", tt_from, tt_to, true, true), SHOULD);
-        transactionBuilder.add(NumericRangeQuery.newLongRange(field + ".ttTo", tt_from, tt_to, true, true),SHOULD);
+        BooleanQuery.Builder containsTransactionBuilder = new BooleanQuery.Builder();
+        containsTransactionBuilder.add(newLongRange(field + ".ttFrom", minTime, ttFromTime, true, true), MUST);
+        containsTransactionBuilder.add(newLongRange(field + ".ttTo ", ttToTime, maxTime, true, true), MUST);
+        transactionBuilder.add(containsTransactionBuilder.build(), SHOULD);
 
-        BooleanQuery.Builder containsTransactionBuilder= new BooleanQuery.Builder();
-        containsTransactionBuilder.add(NumericRangeQuery.newLongRange(field+".ttFrom",MIN,tt_from,true,true),MUST);
-        containsTransactionBuilder.add(NumericRangeQuery.newLongRange(field+".ttTo ",tt_to,MAX,true,true),MUST);
-        transactionBuilder.add(containsTransactionBuilder.build(),SHOULD);
-
-        builder.add(validBuilder.build(),MUST);
-        builder.add(transactionBuilder.build(),MUST);
+        builder.add(validBuilder.build(), MUST);
+        builder.add(transactionBuilder.build(), MUST);
 
         Query query = builder.build();
         query.setBoost(boost);
         return query;
+    }
+
+    private static Long parseTime(BitemporalMapper mapper, Long defaultTime, Object value) {
+        return value == null ?
+               new BitemporalDateTime(defaultTime).toDate().getTime() :
+               mapper.parseBitemporalDate(value).toDate().getTime();
     }
 
     /** {@inheritDoc} */